Addressing Rohit's comments
