+++ src/NuGet.Core/NuGet.CommandLine.XPlat/Utility/TableParser.cs
@@ -1,8 +1,10 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using System.Text;
 
 namespace NuGet.CommandLine.XPlat.Utility
 {
